[add & fix] ミッション関係のコミット

+++ src/main/java/com/github/tanokun/tanorpg/game/player/status/Status.java
@@ -2,14 +2,14 @@ package com.github.tanokun.tanorpg.game.player.status;
 
 public class Status {
     private StatusType statusType;
-    private double level;
+    private int level;
 
-    public Status(StatusType statusType, double lvl){this.statusType = statusType; this.level = lvl;}
+    public Status(StatusType statusType, int lvl){this.statusType = statusType; this.level = lvl;}
 
-    public double getLevel() {
+    public int getLevel() {
         return level;
     }
     public StatusType getStatusType() {return statusType;}
-    public void setLevel(double level) {this.level = level;}
-    public void addLevel(double level) {this.level += level;}
+    public void setLevel(int level) {this.level = level;}
+    public void addLevel(int level) {this.level += level;}
 }
